requirements.txt
Django==3.2.8
djangorestframework==3.12.4
finance-datareader==0.9.31
flatbuffers==1.12
future==0.18.2
Mako==1.1.4
Pillow==8.1.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
serializers==0.2.4
psycopg2
psycopg2-binary
